Feat: Add detailed attendance tracking and statistics features
Introduced new classes and methods for enhanced attendance tracking, including filtering by date, section, and gender. Updated the AttendanceService and repository with logic to compute attendance stats, counts, rankings, and date-based filtering for both classes and schools. Refactored related DTOs and tests to integrate these features.

+++ src/main/java/com/edusyspro/api/dto/custom/AttendanceStatusStat.java
@@ -2,9 +2,9 @@ package com.edusyspro.api.dto.custom;
 
 public record AttendanceStatusStat(
         String date,
-        byte present,
-        byte absent,
-        byte late,
-        byte excused
+        Long present,
+        Long absent,
+        Long late,
+        Long excused
 ) {
 }
